+++ HeMaNe.Web/Controllers/TeamsController.cs
@@ -12,27 +12,37 @@ namespace HeMaNe.Web.Controllers
     public class TeamsController : ControllerBase
     {
         private readonly ITeamService _service;
+        private readonly IUserService _userService;
 
-        public TeamsController(ITeamService service)
+        public TeamsController(ITeamService service, IUserService userService)
         {
             _service = service;
+            _userService = userService;
         }
 
         [HttpGet]
         public async Task<IActionResult> Index()
         {
-            return Ok(await _service.GetAsync(ScopedFilter.Manager));
+            return Ok(await _service.GetAsync(this._userService.CurrentScope()));
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> Index(int id)
         {
+            if (!await this._service.HasAccess(id))
+            {
+                return Forbid();
+            }
             return Ok(await _service.GetAsync(id));
         }
 
         [HttpPost]
         public async Task<IActionResult> Save([FromBody] TeamDto dto)
         {
+            if (!await this._service.HasAccess(dto))
+            {
+                return Forbid();
+            }
             await this._service.SaveAsync(dto);
             return Ok();
         }
@@ -40,6 +50,10 @@ namespace HeMaNe.Web.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!await this._service.HasAccess(id))
+            {
+                return Forbid();
+            }
             await this._service.DeleteAsync(id);
             return Ok();
         }
